\name{opMetaTestingInformation}
\alias{opMetaTestingInformation}
\title{Package functions information}
\description{
A reminder of available functions from this package, and, most common usage
intent. A poor man CLI cheat sheet.
}
\usage{
opMetaTestingInformation()
}
\value{
See \bold{\emph{\code{\link[wyz.code.offensiveProgramming]{opInformation}}}} value
description.
}
\author{
\packageAuthor{wyz.code.metaTesting}

Maintainer: \packageMaintainer{wyz.code.metaTesting}
}
\seealso{
Refer also to package vignettes.
}
\examples{
##---- typical case ----
opMetaTestingInformation()
}
\keyword{ programming }
\keyword{ utilities }
\concept{ meta testing }
